pyproject.toml
[project]
name = "siamese-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "pytorch-lightning>=2.6.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
